requirements.txt
streamlit
openai
python-dotenv
gradio
textblob
requests
beautifulsoup4
pandas